+++ engine/rendering/storage/VertexBuffer.h
@@ -5,13 +5,13 @@
 class VertexBuffer
 {
 public:
-	VertexBuffer(const void* data=nullptr, unsigned int size=36000, bool dynamic=1);
+	VertexBuffer(const void* data=nullptr, size_t size=36000, bool dynamic=1);
 	~VertexBuffer();
 
 	void Bind() const;
 	void Unbind() const;
 
-	void Update(const void* data, unsigned int size);
+	void Update(const void* data, size_t size);
 private:
 	unsigned int m_buffer;
 };
